requirements.txt
requests
beautifulsoup4
lxml
pandas
slackclient
